requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2019.9.11=hecc5488_0
cairo=1.14.12=h8948797_3
certifi=2019.9.11=py36_0
chardet=3.0.4=pypi_0
django=2.2.4=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.56.2=hd408876_0
icu=58.2=h9c2bf20_1
idna=2.8=pypi_0
intel-openmp=2019.4=243
jpeg=9b=h024ee3a_2
libboost=1.67.0=h46d08c1_4
libedit=3.1.20170329=hf8c457e_1001
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
mkl=2019.4=243
mkl_fft=1.0.12=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
molvs=0.1.1=py_0
ncurses=6.1=hf484d3e_1002
numpy=1.16.4=py36h7e9f1db_0
numpy-base=1.16.4=py36hde5b4d6_0
olefile=0.46=py36_0
openssl=1.1.1c=h516909a_0
pandas=0.25.0=py36he6710b0_0
pcre=8.43=he6710b0_0
pillow=6.1.0=py36h34e0f95_0
pip=19.1.1=py36_0
pixman=0.38.0=h7b6447c_0
psycopg2-binary=2.8.3=pypi_0
py-boost=1.67.0=py36h04863e7_4
pypdf2=1.26.0=pypi_0
python=3.6.8=h0371630_0
python-dateutil=2.8.0=py36_0
pytz=2019.1=py_0
rdkit=2019.03.3.0=py36hc20afe1_1
readline=7.0=ha6073c6_4
requests=2.22.0=pypi_0
rope=0.14.0=py_0
setuptools=41.0.1=py36_0
six=1.12.0=py36_0
sqlite=3.28.0=h8b20d00_0
sqlparse=0.3.0=pypi_0
tk=8.6.8=hbc83047_0
urllib3=1.25.3=pypi_0
wheel=0.33.4=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0




# web
Django==2.2.4
numpy==1.16.4
pandas==0.25
  xlrd==1.0.0
  openpyxl==2.4.7
psycopg2==2.7.4
# rdkit==2017.09.3  # install via conda not pip
# git+http://github.com/rdkit/django-rdkit@30cac1f61a1d401a5020a5e410a648fd4b2bfbda
cython==0.28.2
jinja2==2.9.5
scipy==0.19.0
pyRserve==0.9.1
gunicorn==20.0.4
molvs==0.1.1
python-decouple==3.3
# pypdf2==1.26.0
# ca-certificates==2019.9.11
# ncurses==6.1
# psycopg2-binary==2.8.3
# requests==2.22.0
# sqlparse==0.3.0
# idna==2.8
# openssl==1.1.1c
# molvs==0.1.1
# django==2.2.4
# rdkit==2019.03.3.0
# libedit==3.1.20170329
# urllib3==1.25.3
# sqlite==3.28.0
# chardet==3.0.4
